fixed a bug with todos changed visibility for a method moved two services
