+++ src/main/java/de/tum/www1/orion/connector/ide/exercise/OrionExerciseConnector.kt
@@ -7,7 +7,7 @@ import com.intellij.ui.jcef.JBCefJSQuery
 import de.tum.www1.orion.connector.ide.OrionConnector
 import de.tum.www1.orion.dto.Feedback
 import de.tum.www1.orion.dto.ProgrammingExercise
-import de.tum.www1.orion.exercise.OrionAssessmentService
+import de.tum.www1.orion.exercise.assessment.OrionAssessmentService
 import de.tum.www1.orion.exercise.OrionExerciseService
 import de.tum.www1.orion.exercise.OrionFeedbackService
 import de.tum.www1.orion.messaging.OrionIntellijStateNotifier
